Perf: update throttling
we only update if there is a change in the buff list. or the timer is below 20 seconds for buffs

otherwise we will update every 2 min to check-in or be removed as stale after 3 min

/mybuffs mailbox
will show you the message frequency in a table. 
the data is only recorded when you have this window open and clears when you close it.